feat: extract router error-boundaries into dedicated file
