+++ projects/services/kubestro-core/frontend/src/routes.tsx
@@ -1,4 +1,4 @@
-import { createBrowserRouter, isRouteErrorResponse, useRouteError } from 'react-router'
+import { createBrowserRouter } from 'react-router'
 import type { LazyRouteFunction, PatchRoutesOnNavigationFunction, RouteObject, unstable_RouterContextProvider } from 'react-router'
 import { requireSetupMiddleware } from './middlewares/requireSetup'
 import { indexRedirectMiddleware } from './features/redirect/redirect'
@@ -8,6 +8,7 @@ import oidcCallbackPage from './features/authentication/oidc-callback'
 import logoutPage from './features/authentication/logout'
 import { dashboardRoutes } from './features/dashboard/_routes'
 import { federation } from './utils/module-federation'
+import { ErrorBoundary } from './error-boundaries'
 
 declare module 'react-router' {
   interface Future {
@@ -46,41 +47,6 @@ const patchRoutesOnNavigation: PatchRoutesOnNavigationFunction = async ({ path, 
   }
 }
 
-export function ErrorBoundary() {
-  const error = useRouteError()
-
-  let message = 'Oops!'
-  let details = 'An unexpected error occurred.'
-  let stack: string | undefined
-
-  if (isRouteErrorResponse(error)) {
-    message = error.status === 404 ? '404' : 'Error'
-    details =
-      error.status === 404 ?
-        'The requested page could not be found.' :
-        error.statusText || details
-  }
-  else if (import.meta.env.DEV && error && error instanceof Error) {
-    details = error.message
-    // eslint-disable-next-line @typescript-eslint/prefer-destructuring -- destructuring isn't possible here
-    stack = error.stack
-  }
-
-  return (
-    <main className="container mx-auto p-4 pt-16">
-      <h1>{message}</h1>
-      <p>{details}</p>
-
-      {stack ?
-        (
-          <pre className="w-full overflow-x-auto p-4">
-            <code>{stack}</code>
-          </pre>
-        ) :
-        null}
-    </main>
-  )
-}
 export const router = createBrowserRouter([
   {
     id: 'root',
